chore: move title logic into its own file
